% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiralPlot.R
\name{laser_locus}
\alias{laser_locus}
\title{plot a gif}
\usage{
laser_locus(point_locus, parameter, point = NULL)
}
\arguments{
\item{point_locus}{point route}

\item{parameter}{super parameter}

\item{point}{point need to scan}
}
\value{
a gif show the process
}
\description{
plot a gif
}
\examples{
fou = matrix(c(-1000, -1000, 1000, 1000), 2, 2, TRUE)
parameter <- list()
parameter$light_source = c(-1000, -1000)
parameter$FOU = fou
parameter$unit = 'μrad'
parameter$'scanning-time(ms)/step'= 120
parameter$'beam-divergence-angle' = 60
parameter$'compensate-angle' = 5
# 生成扫描点
point_locus = spiral_scan_lines(parameter$'beam-divergence-angle', parameter$'compensate-angle', fou)
# 画gif图，耗时长
laser_locus(point_locus, parameter)

point_locus = spiral_scan_lines(parameter$'beam-divergence-angle', parameter$'compensate-angle', fou, c(500, 500, 30))
laser_locus(point_locus, parameter, c(500, 500, 30))
}
